16. Exercise HTTP Server/http-server-handler-middleware-cp-v1/main.go
package main

import (
	"fmt"
	"net/http"
)

func StudentHandler() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
		w.Write([]byte("Welcome to Student page"))
	}
}

func RequestMethodGet(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if r.Method != "GET" {
			w.WriteHeader(405)
			w.Write([]byte("Method is not allowed"))
			return
		}

		next.ServeHTTP(w, r)
	}) // TODO: replace this
}

func main() {
	// TODO: answer here
	mux := http.DefaultServeMux

	mux.Handle("/student", RequestMethodGet(StudentHandler()))

	fmt.Println("Server running at localhost:8080")
	http.ListenAndServe("localhost:8080", nil)
}
